+++ kafka/integration/integration_test.go
@@ -2211,15 +2211,6 @@ func (its *IntegrationTestSuite) TestAdminACLs() {
 			PermissionType:      ACLPermissionTypeAllow,
 		},
 		{
-			Type:                ResourceTransactionalID,
-			Name:                group,
-			ResourcePatternType: ResourcePatternTypePrefixed,
-			Principal:           "User:test-user-2",
-			Host:                "*",
-			Operation:           ACLOperationAll,
-			PermissionType:      ACLPermissionTypeAllow,
-		},
-		{
 			Type:                ResourceTopic,
 			Name:                topic,
 			ResourcePatternType: ResourcePatternTypePrefixed,
@@ -2230,6 +2221,15 @@ func (its *IntegrationTestSuite) TestAdminACLs() {
 		},
 		{
 			Type:                ResourceGroup,
+			Name:                group,
+			ResourcePatternType: ResourcePatternTypePrefixed,
+			Principal:           "User:test-user-2",
+			Host:                "*",
+			Operation:           ACLOperationAll,
+			PermissionType:      ACLPermissionTypeAllow,
+		},
+		{
+			Type:                ResourceTransactionalID,
 			Name:                group,
 			ResourcePatternType: ResourcePatternTypePrefixed,
 			Principal:           "User:test-user-2",
@@ -2314,7 +2314,7 @@ func (its *IntegrationTestSuite) TestAdminACLs() {
 			resultCreateACLs[0].Error.Code())
 	}
 
-	// DescribeACLs must return the three ACLs
+	// DescribeACLs must return the four ACLs
 	ctx, cancel = context.WithTimeout(context.Background(), maxDuration)
 	defer cancel()
 	resultDescribeACLs, err := a.DescribeACLs(ctx, aclBindingFilters[0], SetAdminRequestTimeout(requestTimeout))
@@ -2335,7 +2335,7 @@ func (its *IntegrationTestSuite) TestAdminACLs() {
 	expectedDeleteACLs = []DeleteACLsResult{
 		{
 			Error:       noError,
-			ACLBindings: newACLs[1:3],
+			ACLBindings: newACLs[1:4],
 		},
 	}
 	if err != nil {