+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Confluent's Golang client for Apache Kafka
 
-## v2.x.x
+## v2.15.0
+
+This is a feature release:
 
 ### Enhancements
-* Add support for `ResourceTransactionalID` in AdminClient ACL APIs (#1407)
+* Add support for `ResourceTransactionalID` in AdminClient ACL APIs (#1407, @jzvelc)
 
 ### Fixes
 * Handle non-HTTP errors during retries (#1563)